webinars/Koyeb/application/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
jinja2
python-dotenv
ollama